General cleanup in kdf+kwikdisk+kcmdf
Usagebars in kdf should now work. There is a problem when
the column size is changed.

kwikdisk segfaults when started from kicker.  In KDE-1.1x
it is not placed in kpanels's tray but rather as a window in the
upper left (and it does not segfault).

kcmdf is no more based on Kcontrol.

svn path=/trunk/kdeutils/kdf/; revision=35906