GUI: reordered UI, added some QWhatsThis, added Ctrl-Shift-Save for quicksaving.
Rational: there are primarily two use cases for ksnapshot:

Common case: a user takes a snapshot, wishes to save it. Having multiple
buttons for saving, esp one that silently saves with no easily visible
feedback as to what has happened, is not very user-friendly. It also
clutters the UI unecessarily. So now there is just one button called
Save (not Save as..., since it is linked to the Save shortcut). This
behaviour is also functionally consistent with other KDE apps in that
it is saving a new file; it has not openned an existing file and then
subsequently resaving it. While there is a default filename provided,
this is different than an existing openned file.

Power Use Case: a (typically advanced) user takes several snapshots in
succession and wants to quickly save them. There is now a "backdoor"
shortcut (noted in the What's This help) that enables this quick saving:
Ctrl+Shift+S. In combination with the autoincrementing filename feature,
a series of snapshots can be taken in quick succession.

svn path=/trunk/kdegraphics/ksnapshot/; revision=242774