KSnaphot uses KWin's new screenshot effect for grabbing windows.
This allows to capture a window including it's shadow and translucency.
The KWin effect is invoked via DBus and KWin passes a pixmap handle of
the screenshot to KSnapshot. If the screenshot effect is not available
the normal mode is used.

Thanks to Aaron for review.

Please see http://reviewboard.kde.org/r/4814/


svn path=/trunk/KDE/kdegraphics/ksnapshot/; revision=1175357